Improve dictionary description
I improved dictionary description to ad some real life examples and expand on it a little bit so person who never seen dictionary can understand it better seeing them for a first time here. 
I also added note about the way to compare dictionaries  as it can catch people of guard because common assumption is that it works the same as comparing Arrays when it does not.