Revert "Enable the ability to use Godot as a subview within an Android app"
This reverts commit 920639511d5e774402c6b4f37ee0550368f5109a.

The changes are good, this revert is only done for release management reasons
as we want this feature to get more testing before making it in a stable build,
but a 3.2.3 release is imminent to handle some regressions in 3.2.2.

This will be re-committed in a 3.2-based feature branch, and we'll merge it
again once we're confident about it (probably for 3.2.4).